% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/baf_plot.R
\name{baf_plot}
\alias{baf_plot}
\title{baf_plot}
\usage{
baf_plot(count_df)
}
\arguments{
\item{count_df}{a data.frame containing info}
}
\description{
BAlleleFreq plot.
}
\references{
Laurie, Cathy C, Kimberly F Doheny, Daniel B Mirel, Elizabeth W Pugh, Laura J Bierut, Tushar Bhangale, Frederick Boehm, Neil E Caporaso, Marilyn C Cornelis, Howard J Edenberg and others.
  2010. Quality control and quality assurance in genotypic data for genome-wide association studies.
  Genetic Epidemiology 34(6): 591--602.
}

