% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{baf_stats}
\alias{baf_stats}
\title{baf_stats}
\usage{
baf_stats(inMatrix, minq = 0L)
}
\arguments{
\item{inMatrix}{input matrix}

\item{minq}{minimum quality for call to be retained}
}
\description{
Convert mpileup output to count tables
}
\details{
The character matrix \strong{inMatrix} should consist of columns  5 columns. 
The first column is the chromosome name (and is not presently used).
The second column is the chromosomal position.
The third column is the reference allele.
The fourth column is a string of calls.
The fifth column is a string of qualities.
This is expected to come from mpileup output.
Note that while mpileup can include data for multiple samples, here we need to process each sample seperately.

The reference alleles must be in all upper case.
See \code{toupper} if they are not.
}

