% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/baf_summary.R
\name{baf_summary}
\alias{baf_summary}
\title{Summarize baf count tables}
\usage{
baf_summary(counts, probs = c(0.025, 0.16, 0.5, 0.84, 0.975))
}
\arguments{
\item{counts}{count table produced by baf_stats}

\item{probs}{vector of probabilities to pass to quantile}
}
\value{
a vector of size two plus the number of quantiles requested.
}
\description{
Calculate summary statistics for count tables produced by baf_stats
}
\details{
This function is intended to summarize count tables produced by baf_stats.
By default the mean, variance and five quantiles (0.025, 0.16, 0.5, 0.84, 0.975) are returned.
The quantiles are generated by the function quantile and may be changed by changing the option probs.
}
\seealso{
baf_stats, quantile
}

