\name{covR-package}
\alias{covR-package}
\alias{covR}
\docType{package}
\title{
\packageTitle{covR}
}
\description{
\packageDescription{covR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{covR}
\packageIndices{covR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{covR}

Maintainer: \packageMaintainer{covR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
