% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/covR.R
\docType{package}
\name{covR}
\alias{covR}
\alias{covR-package}
\title{covR: A package for exploring sequence coverage.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

