% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/file_io.R
\name{file_io}
\alias{file_io}
\alias{file_stats}
\title{file_io}
\usage{
file_stats(filename, sep = "\\t", skip = 0L)
}
\arguments{
\item{filename}{name os file containing tabular data}

\item{sep}{character which delimits columns}

\item{skip}{number of rows to skip before reading data}
}
\description{
Tools for rapid input of tabular data.
}
\details{


The function file_stats scans a file and returns the number of columns in the first row and the total number of rows.
}

