% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pie_exposure.R
\name{plot_pie_exposure}
\alias{plot_pie_exposure}
\title{plot pie of human exposure}
\usage{
plot_pie_exposure(man = NULL, cow = NULL, indoor = NULL,
  intervene_indoor = 0, intervene_cow = 0, intervene_outdoor = 0,
  col = c("deepskyblue", "orange", "yellow"))
}
\arguments{
\item{man}{proportion of indoor mosquitoes feeding on man}

\item{cow}{proportion of outdoor mosquitoes feeding on livestock}

\item{indoor}{proportion of mosquitoes feeding indoors}

\item{intervene_indoor}{proportion coverage of intervention targetting indoor}

\item{intervene_cow}{proportion coverage of intervention targetting livestock}

\item{intervene_outdoor}{proportion coverage of intervention targetting outdoor}
}
\value{
nothing
}
\description{
plots human exposure based on behaviour inputs
}
\examples{
plot_pie_exposure(man=0.5, indoor=0.7, intervene_indoor=0.5)

}

