% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pie_feeding.R
\name{plot_pie_feeding}
\alias{plot_pie_feeding}
\title{plot pie of feeding}
\usage{
plot_pie_feeding(man = NULL, cow = NULL, indoor = NULL,
  intervene_indoor = 0.8, intervene_cow = 0, intervene_outdoor = 0.2,
  col = c("deepskyblue", "orange", "yellow"))
}
\arguments{
\item{man}{proportion of indoor mosquitoes feeding on man}

\item{cow}{proportion of outdoor mosquitoes feeding on livestock}

\item{indoor}{proportion of mosquitoes feeding indoors}

\item{intervene_indoor}{proportion coverage of intervention targetting indoor}

\item{intervene_cow}{proportion coverage of intervention targetting livestock}

\item{intervene_outdoor}{proportion coverage of intervention targetting outdoor}
}
\value{
dataframe of some plot coords
}
\description{
plots vector feeding based on behaviour inputs
}
\examples{
plot_pie_feeding(man=0.5, indoor=0.7, intervene_indoor=0.5)

}

