% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{File input/output}
\alias{File input/output}
\alias{file_stats}
\alias{read_matrix}
\title{File input/output}
\usage{
file_stats(filename, sep = "\\t", nrows = -1L, skip = 0L, verbose = 1L)

read_matrix(filename, sep = "\\t", nrows = 1L, cols = 0L, skip = 0L,
  verbose = 1L)
}
\arguments{
\item{filename}{name of a file}

\item{sep}{character which delimits columns}

\item{nrows}{number of rows to read}

\item{skip}{number of rows to skip}

\item{verbose}{should verbose output be generated}

\item{cols}{vector of column numbers to include in the matrix}
}
\value{
\strong{file_stats} returns a three element vector.
'Total_rows' reports the total number of rows read.
 This is either the number of rows in the file or the number of skipped rows and the number of rows read in.
 'Rows' is the number of rows read in.
 This is either the same as nrows or however many rows were read in after skip and before the end of the file (when less than nrows).

\strong{read_matrix} returns a matrix of strings of dimension specified by nrows and ncols.
}
\description{
Fast but featureless input of tabular data in either *.txt or *.gz format.
}
\seealso{
\href{http://cran.r-project.org/web/packages/readr/index.html}{readr}
\href{http://cran.r-project.org/web/packages/data.table/index.html}{data.table::fread}
}

