% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUI1.r
\name{runUI1}
\alias{runUI1}
\title{run an interactive shiny User Interface to the graphical model}
\usage{
runUI1()
}
\value{
nothing
}
\description{
displays a user interface in the browser, allowing users to modify inputs
}

