% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{write_matrix}
\alias{write_matrix}
\title{write_matrix}
\usage{
write_matrix(filename, mymatrix, sep = "\\t", verbose = 1L)
}
\arguments{
\item{filename}{filename for output}

\item{mymatrix}{matrix to be written to file}

\item{sep}{delimiting character}

\item{verbose}{should verbose output be generated?}
}
\description{
write_matrix
}

