% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baf_plot.R
\name{baf_plot}
\alias{baf_plot}
\title{baf_plot}
\usage{
baf_plot(counts, alpha = 255, title = "Locus", na.rm = FALSE, ...)
}
\arguments{
\item{counts}{a numeric matrix containing count data}

\item{alpha}{opacity (0-255)}

\item{title}{for the plot.}

\item{na.rm}{Logical, should sites which contain NAs (no reported coverage) be removed}

\item{...}{arguments to be passed to methods}
}
\description{
B allele frequency plot.
}
\details{
Plots coverage and allele frequencies for a chromosomal region.
The upper pane presents per site sequencing depth while the lower pane presents the frequency of all four alleles.
These plots can be used to identify technical error (i.e., read alignment issues) and/or biological phenomena such as copy number variation.

The matrix of count data can be generated from mpileup data with the function \code{\link{baf_stats}}
}
\references{
Laurie, Cathy C, Kimberly F Doheny, Daniel B Mirel, Elizabeth W Pugh, Laura J Bierut, Tushar Bhangale, Frederick Boehm, Neil E Caporaso, Marilyn C Cornelis, Howard J Edenberg and others.
  2010. Quality control and quality assurance in genotypic data for genome-wide association studies.
  Genetic Epidemiology 34(6): 591--602.
}
\seealso{
\code{\link{baf_stats}}
}

