% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/baf_plot.R
\name{baf_plot}
\alias{baf_plot}
\title{baf_plot}
\usage{
baf_plot(counts, alpha = 255, title = "Locus", na.rm = FALSE, ...)
}
\arguments{
\item{counts}{a numeric matrix containing count data}

\item{alpha}{opacity (0-255)}

\item{title}{for the plot.}

\item{na.rm}{Logical, should sites which contain NAs be removed}

\item{...}{arguments to be passed to methods}
}
\description{
BAlleleFreq plot.
}
\references{
Laurie, Cathy C, Kimberly F Doheny, Daniel B Mirel, Elizabeth W Pugh, Laura J Bierut, Tushar Bhangale, Frederick Boehm, Neil E Caporaso, Marilyn C Cornelis, Howard J Edenberg and others.
  2010. Quality control and quality assurance in genotypic data for genome-wide association studies.
  Genetic Epidemiology 34(6): 591--602.
}

