% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{baf_stats}
\alias{baf_stats}
\title{baf_stats}
\usage{
baf_stats(calls, quals, ref, minq)
}
\arguments{
\item{calls}{vector of pileup calls}

\item{quals}{vector of pileup calls}

\item{ref}{vector of reference alleles}

\item{minq}{minimum quality for call to be retained}
}
\description{
baf_stats
}

