% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{baf_stats}
\alias{baf_stats}
\alias{baf_stats_st}
\title{baf_stats}
\usage{
baf_stats(calls, quals, ref, minq = 0L)

baf_stats_st(inMatrix, minq = 0L)
}
\arguments{
\item{calls}{vector of pileup calls}

\item{quals}{vector of pileup calls}

\item{ref}{vector of reference alleles}

\item{minq}{minimum quality for call to be retained}

\item{inMatrix}{input matrix}
}
\description{
baf_stats

baf_stats_st
}
\details{
The reference alleles must be in all upper case.
See \code{toupper} if they are not.

The character matrix 'inMatrix' should consist of columns for position, reference allele, calls and qualities.
This is expected to come from mpileup output.
Note that while mpileup can include data for multiple samples, here we need to process each sample seperately.
}

