% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bedify}
\alias{bedify}
\title{Parse data by a bed file}
\usage{
bedify(bed, mydata)
}
\arguments{
\item{bed}{data.frame}

\item{mydata}{StringMatrix to be sorted}
}
\description{
seperate a data matrix by using bed format data.
}

