% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bedify}
\alias{bedify}
\title{Parse data by a bed file}
\usage{
bedify(myBed, myData)
}
\arguments{
\item{myBed}{matrix of bed format data}

\item{myData}{StringMatrix or IntegerMatrix to be sorted}
}
\description{
Seperate a data matrix into list elements based on coordinates from bed format data.
}
\details{
\strong{Bed format} data contain at least four columns.
The first column indicates the chromosome (i.e., supercontig, scaffold, contig, etc.).
The second cotains the starting positions.
The third the ending positions.
The fourth are the names of the features.
All subsequent columns are ignored here.
In an attempt to optimize performance the data are expected to be formatted as a character matrix.
The starting and end positions are converted to numerics internally.

The \strong{matrix format} used here is based on vcf type data.
Typically these data have a chromosome as the first column.
Each chromosome has its own coordinate system which begins at one.
This means that using multiple chromosomes will necessitate some fix to the coordinate systems.
Here I take the perspective that you should simply work on one chromosome at a time, so the chromosome information is ignored.
The first column is the chromosome, which I ignore.
The second column is the position, which is used for sorting.
Subsequent columns are not treated but are brought along with the subset.


When the matrix is of numeric form the first column, which contains the chromosome identifier (CHROM), must also be numeric.
This is because matrix elements must all be of the same type.



\href{https://genome.ucsc.edu/FAQ/FAQformat.html#format1}{Bed format} at UCSC
}

