% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bedify}
\alias{bedify}
\alias{bedify_sm}
\title{Parse data by a bed file}
\usage{
bedify(myBed, myData)

bedify_sm(myBed, myData)
}
\arguments{
\item{myBed}{matrix of bed format data}

\item{myData}{StringMatrix to be sorted}
}
\description{
seperate a data matrix by using bed format data.

Parse data by a bed file
}
\details{
Bed format data contain at least four columns.
The first column indicates the chromosome (i.e., supercontig, scaffold, contig, etc.).
The second cotains the starting positions.
The third the ending positions.
The fourth are the names of the features.
All subsequent columns are ignored here.
In an attempt to optimize performance the data are expected to be formatted as a character matrix.
The starting and end positions are converted to numerics internally.


\href{https://genome.ucsc.edu/FAQ/FAQformat.html#format1}{Bed format} at UCSC
}

