% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{File input/output}
\alias{File input/output}
\alias{file_stats}
\alias{read_matrix}
\title{File input/output}
\usage{
file_stats(filename, sep = "\\t", nrows = -1L, skip = 0L, verbose = 1L)

read_matrix(filename, sep = "\\t", nrows = 1L, ncols = 1L, skip = 0L,
  verbose = 1L)
}
\arguments{
\item{filename}{name of a file}

\item{sep}{character which delimits columns}

\item{nrows}{number of rows to read}

\item{skip}{number of rows to skip}

\item{verbose}{should verbose output be generated}

\item{ncols}{number of columns for the matrix}
}
\description{
File input/output
}
\seealso{
\href{http://cran.r-project.org/web/packages/readr/index.html}{readr}
\href{http://cran.r-project.org/web/packages/data.table/index.html}{data.table::fread}
}

