% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{File input}
\alias{File input}
\alias{file_stats}
\alias{read_matrix}
\title{File input}
\usage{
file_stats(filename, sep = "\\t", nrows = -1L, skip = 0L, verbose = 1L)

read_matrix(filename, sep = "\\t", nrows = 1L, cols = 0L, skip = 0L,
  verbose = 1L)
}
\arguments{
\item{filename}{name of a file}

\item{sep}{character which delimits columns}

\item{nrows}{number of rows to read}

\item{skip}{number of rows to skip}

\item{verbose}{should verbose output be generated}

\item{cols}{vector of column numbers to include in the matrix}
}
\value{
\strong{file_stats} returns a three element vector.

\strong{read_matrix} returns a matrix of strings
}
\description{
Fast but featureless input of tabular data in either *.txt or *.gz format.
}
\details{
\strong{file_stats} returns a three element vector containing a summary of a file's contents.
'Total_rows' reports the total number of rows read.
 This is either the number of rows in the file or the number of skipped rows and the number of rows read in.
 'Rows' is the number of rows read in.
 This is either the same as nrows or however many rows were read in after skip and before the end of the file (when less than nrows).
 'Columns' is the number of columns resulting after delimiting with sep.
 This information is intended to be used with read_matrix.

\strong{read_matrix} returns a matrix of strings of dimension specified by nrows and cols.
The cols parameter is a vector of integers specifying which columns to read in.
}
\seealso{
\href{http://cran.r-project.org/package=readr}{readr}
\href{http://cran.r-project.org/package=data.table}{data.table::fread}
}

