% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{write_matrix}
\alias{write_matrix}
\title{write_matrix}
\usage{
write_matrix(filename, mymatrix, sep = "\\t", verbose = 1L)
}
\arguments{
\item{filename}{filename for output}

\item{mymatrix}{matrix to be written to file}

\item{sep}{delimiting character}

\item{verbose}{should verbose output be generated?}
}
\description{
Write matrix data to a gzipped file.
}
\details{
Writes matrix data to a gzipped file delimited by sep.
Data is appended to the file.
This is intended to allow header information to be included by other calls.
It should also allow files to grow with incremental processes.
}

